+++ src/store/modules/post.js
@@ -3,21 +3,27 @@ import * as types from '../mutation-types'
 
 // initial state
 const state = {
-  all: [],
   recent: [],
   loaded: false
 }
 
 // getters
 const getters = {
-  allPosts: state => state.all,
-  allPostsLoaded: state => state.loaded
+  recentPosts: state => limit => {
+    if ( !limit || !_.isNumber(limit) || _.isNull(limit) || typeof limit == 'undefined' ) {
+      return state.recent
+    }
+    let recent = state.recent
+    return recent.slice(0, limit)
+  },
+
+  recentPostsLoaded: state => state.loaded
 }
 
 // actions
 const actions = {
-  getAllPosts ({ commit }) {
-    api.getPosts(posts => {
+  getPosts ({ commit }, { limit }) {
+    api.getPosts(limit, posts => {
       commit(types.STORE_FETCHED_POSTS, { posts })
       commit(types.POSTS_LOADED, true)
       commit(types.INCREMENT_LOADING_PROGRESS)
@@ -28,7 +34,7 @@ const actions = {
 // mutations
 const mutations = {
   [types.STORE_FETCHED_POSTS] (state, { posts }) {
-    state.all = posts
+    state.recent = posts
   },
 
   [types.POSTS_LOADED] (state, val) {