+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - More documnentation. @EvanAgee
 - Post module. @EvanAgee
 - A few code comments. @EvanAgee
+- Example widgets for posts and pages. @EvanAgee
+
+## Changed
+- Home page to display the posts and pages widgets. @EvanAgee
 
 ## 2017-10-11
 Initial commit